backend/requirements.txt
# Core web framework
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6

# AI and ML libraries
openai==1.3.5
transformers==4.35.2
torch>=2.0.0
numpy==1.24.3
opencv-python==4.8.1.78
mediapipe==0.10.8

# Voice processing
openai-whisper==20231117
elevenlabs==0.2.26
pydub==0.25.1
soundfile==0.12.1
pyaudio==0.2.11

# Data handling and storage
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30

# Utilities
python-dateutil==2.8.2
pillow==10.1.0

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
